requirements.txt
torch==2.8.0
torchvision==0.22.0
pillow==11.3.0
numpy==2.2.6
matplotlib==3.10.3
scikit-learn==1.7.1
opencv-python==4.11.0.86
segment-anything==1.0